Move Glue v1 converters into Glue package
